requirements.txt
pandas==2.2.2
panflute==2.3.1
PyYAML==6.0.1
PyYAML==6.0.1
selenium==4.20.0
